add out files

+++ rust/examples/tuples/define-types/src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
-    let things: (i32, char, String) = (11, '2', String::from("three"));
+    let things: (i8, char, String) = (11, '2', String::from("three"));
     println!("{:?}", things);
 }
